requirements.txt
python-telegram-bot==13.15
APScheduler==3.9.1
requests==2.31.0
pymongo==4.4.0
python-dotenv==1.0.0
SpeechRecognition==3.8.1
pydub==0.25.1
ffmpeg-python==0.6.1
gunicorn==20.1.0
PyYAML==6.0
